TDoc Digest Deployment Implementation
TDoc Digest is an interactive web application for generating abstractive summary of a 3GPP TDoc contribution. The tool takes the meeting id and TDoc number as inputs and processes the content of the TDoc, i.e., the long original text and produces an abstractive summary  of the TDoc along with a semantic score. The web API application is implemented using the Flask micro web framework. The code is given here along with some guidance on how to set up developer environment and run the web API.